Split databases SDK into files
